+++ examples/experimental/assets_yaml_dsl/assets_yaml_dsl/__init__.py
@@ -1,8 +1,8 @@
 from dagster import Definitions
 from dagster._core.external_execution.subprocess import SubprocessExecutionResource
 
-from .assets_dsl import get_asset_dsl_example_defs
-from .stocks_dsl import get_stocks_dsl_example_defs
+from .pure_assets_dsl.assets_dsl import get_asset_dsl_example_defs
+from .domain_specific_dsl.stocks_dsl import get_stocks_dsl_example_defs
 
 defs = Definitions(
     assets=get_asset_dsl_example_defs() + get_stocks_dsl_example_defs(),
