+++ examples/experimental/assets_yaml_dsl/assets_yaml_dsl/domain_specific_dsl/stocks.yaml
