+++ examples/experimental/assets_yaml_dsl/assets_yaml_dsl/pure_assets_dsl/assets.yaml
