+++ examples/experimental/assets_yaml_dsl/assets_yaml_dsl_tests/test_stocks_dsl.py
@@ -15,10 +15,10 @@ forecast:
 
 
 import yaml
-from assets_yaml_dsl.stocks_dsl import assets_defs_from_stock_assets
+from assets_yaml_dsl.domain_specific_dsl.stocks_dsl import assets_defs_from_stock_assets
 from dagster import AssetKey
 from dagster._core.definitions import materialize
-from examples.experimental.assets_yaml_dsl.assets_yaml_dsl.stocks_dsl import (
+from examples.experimental.assets_yaml_dsl.assets_yaml_dsl.domain_specific_dsl.stocks_dsl import (
     build_stock_assets_object,
 )
